Develop

+++ db/migrate/20190816141521_delete_events_on_gp_calendar_events.rb
@@ -1,6 +1,7 @@
 class DeleteEventsOnGpCalendarEvents < ActiveRecord::Migration[5.0]
   def up
     GpCalendar::Event.all.each do |event|
+      next if event.content.blank?
       event.destroy if event.periods.blank?
     end
   end
